pin XSPRESS3 to commit 0e8eb4ba
While setting up an Xspress3 IOC from the master branch at TES it was discovered that an extra frame was acquired upon setting the acquire PV when in TTL_VETO_ONLY mode. This issue was not observed at SRX which was using the older commit 0e8eb4ba. This commit worked correctly at TES and so we are pinning to it until the problem is resolved on the master branch.